Moved the blocks to blocks.py
